made unit test even easier, minor changes to file naming
